requirements.txt
tensorflow
numpy
matplotlib.pyplot
keras